induction principle for identity types now based
Switched the induction principle for identity types to the default induction principle in Egbert Rijke's textbook and what is called "based path induction" in the UFP textbook. This allows for the elimination and computation rules of identity types to be expressed without the need for dependent function types, allowing for a presentation of the type theory in the standard natural deduction manner, only one where identity types come before every other type.